+++ modules/mvstore/src/main/scala/org/ergoplatform/uexplorer/mvstore/kryo/KryoSerialization.scala
@@ -22,24 +22,6 @@ object KryoSerialization {
     implicit val valueByBoxCodec: DbCodec[java.util.Map[BoxId, Value]] = ValueByBoxCodec
     implicit val blockMetadataCodec: DbCodec[BlockInfo]                = BlockInfoCodec
     implicit val addressCodec: DbCodec[Address]                        = AddressCodec
-
-    def javaSetOf[T](e: T): java.util.Set[T] = {
-      val set = new java.util.HashSet[T]()
-      set.add(e)
-      set
-    }
-    def javaMapOf[K, V](k: K, v: V): java.util.Map[K, V] = {
-      val map = new java.util.HashMap[K, V]()
-      map.put(k, v)
-      map
-    }
-    def javaMapOf[K, V](kvs: IterableOnce[(K, V)]): java.util.Map[K, V] = {
-      val map = new java.util.HashMap[K, V]()
-      kvs.iterator.foreach { case (k, v) =>
-        map.put(k, v)
-      }
-      map
-    }
   }
 
   val pool: Pool[Kryo] = new Pool[Kryo](true, false, 8) {
